requirements.txt
aiofiles==0.6.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.3
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
foreman==0.9.7
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==1.1.0
MarkupSafe==1.1.1
networkx==3.2.1
numpy==1.26.4
otree==5.10.4
packaging==24.0
pandas==2.2.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.5
pytz==2024.1
requests==2.31.0
requests-futures==1.0.1
setuptools==69.2.0
six==1.16.0
SQLAlchemy==1.3.22
sqlparse==0.4.4
starlette==0.14.1
static3==0.7.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.13.4
websockets==10.1
wheel==0.42.0
whitenoise==6.6.0
WTForms==2.3.3
WTForms-SQLAlchemy==0.2

